+++ packages/signer/signer-xrp/src/api/SignerXrp.ts
@@ -1,0 +1,20 @@
+import { type GetAddressDAReturnType } from "@api/app-binder/GetAddressDeviceActionTypes";
+import { type GetAppConfigDAReturnType } from "@api/app-binder/GetAppConfigDeviceActionTypes";
+import { type SignTransactionDAReturnType } from "@api/app-binder/SignTransactionDeviceActionTypes";
+import { type AddressOptions } from "@api/model/AddressOptions";
+import { type TransactionOptions } from "@api/model/TransactionOptions";
+
+export interface SignerXrp {
+  getAppConfig: () => GetAppConfigDAReturnType;
+
+  getAddress: (
+    derivationPath: string,
+    options?: AddressOptions,
+  ) => GetAddressDAReturnType;
+
+  signTransaction: (
+    derivationPath: string,
+    transaction: Uint8Array,
+    options?: TransactionOptions,
+  ) => SignTransactionDAReturnType;
+}