+++ packages/signer/signer-xrp/src/internal/use-cases/transaction/di/transactionModule.ts
@@ -1,0 +1,9 @@
+import { ContainerModule } from "inversify";
+
+import { transactionTypes } from "@internal/use-cases/transaction/di/transactionTypes";
+import { SignTransactionUseCase } from "@internal/use-cases/transaction/SignTransactionUseCase";
+
+export const transactionModuleFactory = () =>
+  new ContainerModule(({ bind }) => {
+    bind(transactionTypes.SignTransactionUseCase).to(SignTransactionUseCase);
+  });