+++ packages/signer/signer-polkadot/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.2.0
+
+### Minor Changes
+
+- [#1530](https://github.com/LedgerHQ/device-sdk-ts/pull/1530) [`dce397b`](https://github.com/LedgerHQ/device-sdk-ts/commit/dce397b44910deb11a531f8b98c205de33d0ae4b) Thanks [@cted-ledger](https://github.com/cted-ledger)! - Implement the Polkadot-family signer APDU layer: GetAddressCommand (with SS58 prefix), multi-chunk SignTransactionCommand (INIT/ADD/LAST), Polkadot application error codes, derivation-path encoding, and SignTransactionTask with blob+metadata chunking
+
+- [#1501](https://github.com/LedgerHQ/device-sdk-ts/pull/1501) [`22cd048`](https://github.com/LedgerHQ/device-sdk-ts/commit/22cd048dfad36ddcf62398e6fdba6f6045702482) Thanks [@cted-ledger](https://github.com/cted-ledger)! - Init polkadot device management kit signer
+
 All notable changes to this project will be documented in this file.
 
 ## [0.1.0] - 2026-05-16